+++ packages/editor-common/web/src/Components/LinkPanel.jsx
@@ -94,7 +94,7 @@ class LinkPanel extends Component {
     return {
       type: 'url',
       className: textInputClassName,
-      placeholder: this.props.t('LinkPanel_InputPlaceholder'),
+      placeholder: this.props.placeholder || this.props.t('LinkPanel_InputPlaceholder'),
       'data-hook': 'linkPanelInput',
       onBlur: () => this.setState({ showValidation: true }),
     };
@@ -172,5 +172,6 @@ LinkPanel.propTypes = {
   dropDown: PropTypes.object,
   onEnter: PropTypes.func,
   onEscape: PropTypes.func,
+  placeholder: PropTypes.string,
 };
 export default LinkPanel;